Allow quotes in identifiers.
The documentation states that the precise specification of identifier syntax
is intentionally vague. Yet I believe that one should be able to call
arbitrary OCaml code from malfunction. For this to be possible, this means
that all OCaml identifiers need to be valid malfunction identifier. As per
the OCaml syntax, this means we have to allow single quotes in malfunction.